% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_elm_header.R
\name{add_elm_header}
\alias{add_elm_header}
\title{Add EML header to a \code{tibble}}
\usage{
add_elm_header(x)
}
\arguments{
\item{x}{Object of class \code{md_tibble}}
}
\description{
This function adds standard EML information to a tibble, such that when
it is converted to xml, it parses correctly as EML. This is not applied
as standard by \code{as_elm_xml()} because it is plausible that some (many?)
users may not wish their files to be interpretable as EML.
}
