% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_eml_row.R
\name{add_eml_row}
\alias{add_eml_row}
\title{Add EML header to an \code{md_tibble} object}
\usage{
add_eml_row(x)
}
\arguments{
\item{x}{Object of class \code{md_tibble}}
}
\description{
This function adds standard EML information to a tibble, such that when
it is converted to xml, it parses correctly as EML. This is not applied
as standard by \code{as_md_xml()} because it is plausible that some (many?)
users may not wish their files to be interpretable as EML.
}
