% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_eml-md.R, R/as_eml-tibble.R, R/as_eml-xml.R,
%   R/as_eml.R
\name{as_eml.md}
\alias{as_eml.md}
\alias{as_eml.tbl_df}
\alias{as_eml.xml_document}
\alias{as_eml}
\title{Convert metadata to class \code{eml}}
\usage{
\method{as_eml}{md}(x)

\method{as_eml}{tbl_df}(x, ...)

\method{as_eml}{xml_document}(x, ...)

as_eml(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{other arguments, currently ignored}
}
\description{
\code{eml} is metadata stored as a list. It is a hack so that we don't have to
overwrite \code{xml2::as_list} or \code{rlang::as_list}. In theory \code{eml} format should
be consistent with the \code{EML} package, but I haven't checked this.
}
