% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_md-tibble.R, R/as_md-eml.R, R/as_md-xml.R,
%   R/as_md.R
\name{as_md.tibble}
\alias{as_md.tibble}
\alias{as_md.eml}
\alias{as_md.xml_document}
\alias{as_md}
\title{Convert metadata to class \code{md}}
\usage{
\method{as_md}{tibble}(x, ...)

\method{as_md}{eml}(x, ...)

\method{as_md}{xml_document}(x, ...)

as_md(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{other arguments, currently ignored}
}
\description{
This is basically a vector
}
