% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_md_list.R
\name{as_md_list}
\alias{as_md_list}
\alias{as_md_list.md_chr}
\alias{as_md_list.md_tibble}
\alias{as_md_list.md_xml}
\title{Convert metadata to class \code{md_list}}
\usage{
as_md_list(x, ...)

\method{as_md_list}{md_chr}(x, ...)

\method{as_md_list}{md_tibble}(x, ...)

\method{as_md_list}{md_xml}(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Other arguments, currently ignored}
}
\description{
\code{md_list} is metadata stored as a list. It is a hack so that we don't have to
overwrite \code{xml2::as_list} or \code{rlang::as_list}. In theory \code{eml} format should
be consistent with the \code{EML} package, but I haven't checked this.
}
