% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_md_tibble.R
\name{as_md_tibble}
\alias{as_md_tibble}
\alias{as_md_tibble.md_chr}
\alias{as_md_tibble.md_list}
\alias{as_md_tibble.md_xml}
\title{Convert metadata to class \code{md_tibble}}
\usage{
as_md_tibble(x, ...)

\method{as_md_tibble}{md_chr}(x, ...)

\method{as_md_tibble}{md_list}(x, ...)

\method{as_md_tibble}{md_xml}(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Other arguments, currently ignored}
}
\description{
This isn't any different from a tibble, but allows us more control over
OOP.
}
