% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble-md.R, R/as_tibble-eml.R,
%   R/as_tibble-xml.R
\name{as_tibble-eml}
\alias{as_tibble-eml}
\alias{as_tibble.md}
\alias{as_tibble.eml}
\alias{as_tibble.xml_document}
\title{Convert metadata to a \code{tibble}}
\usage{
\method{as_tibble}{md}(x, ...)

\method{as_tibble}{eml}(x, ..., .rows, .name_repair, rownames)

\method{as_tibble}{xml_document}(x, ...)
}
\arguments{
\item{x}{An object to convert}

\item{...}{Arguments passed to other methods. Currently ignored.}

\item{.rows}{Currently ignored}

\item{.name_repair}{Currently ignored}

\item{rownames}{Currently ignored}
}
\description{
Placeholder for help files. Actually exported from package \code{tibble}.
}
