% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble-md.R, R/as_tibble-eml.R,
%   R/as_tibble-xml.R, R/as_tibble.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble.md}
\alias{as_tibble.eml}
\alias{as_tibble.xml_document}
\title{convert metadata to a \code{tibble}}
\usage{
\method{as_tibble}{md}(x, ...)

\method{as_tibble}{eml}(x)

\method{as_tibble}{xml_document}(x, ...)
}
\description{
Placeholder for help files. Actually exported from package \code{tibble}.
}
