% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xml-md.R, R/as_xml-eml.R, R/as_xml-tibble.R,
%   R/as_xml.R
\name{as_xml.md}
\alias{as_xml.md}
\alias{as_xml.eml}
\alias{as_xml.tbl_df}
\alias{as_xml}
\title{Convert metadata to class \code{xml}}
\usage{
\method{as_xml}{md}(x)

\method{as_xml}{eml}(x, ...)

\method{as_xml}{tbl_df}(x)

as_xml(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{other arguments, currently ignored}
}
\description{
largely powered by \code{xml2}
}
