% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_elm.R
\name{check_elm}
\alias{check_elm}
\title{Check validity of a metadata statement}
\usage{
check_elm(x, schema = c("xml", "dc", "eml-2.2.0", "eml-gbif-profile-1.3"))
}
\arguments{
\item{x}{An object of class \code{xml_document} from xml2, or optionally a file.}

\item{schema}{Vector of strings selecting one or more schemas to check
against.}
}
\value{
Invisibly returns a tibble showing parsed errors; or an empty
tibble if no errors are identified.
}
\description{
In the Darwin Core standard, metadata statements are mandatory and must be
provided in Ecological Meta Language (EML) in a file called \code{eml.xml}. This
function applies a series of checks designed by GBIF to check the structure
of the specified xml document for consistency with the standard. Note,
however, that this function doesn't check the \emph{content} of those files,
meaning a file could be structurally sound and still be lacking critical
information.
}
\details{
This function uses local versions of \code{dc.xsd}, \code{eml-gbif-profile.xsd} and
\code{eml.xsd} downloaded
from \link{http://rs.gbif.org/schema/eml-gbif-profile/1.3/} on 2024-09-25.
}
