% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eml.R
\name{check_eml}
\alias{check_eml}
\title{Check validity of a metadata statement}
\usage{
check_eml(x, file)
}
\arguments{
\item{x}{Object of any class handled by \code{delma}; i.e. \code{character},
\code{tbl_df}, \code{list} or \code{xml_document}.}

\item{file}{Alternatively an EML file to check. If both \code{x} and \code{file} are
supplied, \code{x} is chosen.}
}
\value{
Invisibly returns a tibble showing parsed errors; or an empty
tibble if no errors are identified.
}
\description{
In the Darwin Core standard, metadata statements are mandatory and must be
provided in Ecological Meta Language (EML) in a file called \code{eml.xml}. This
function applies a series of checks designed by GBIF to check the structure
of the specified xml document for consistency with the standard. Note,
however, that this function doesn't check the \emph{content} of those files,
meaning a file could be structurally sound and still be lacking critical
information.
}
\details{
This function uses local versions of \code{dc.xsd}, \code{eml-gbif-profile.xsd} and
\code{eml.xsd} downloaded
from \code{http://rs.gbif.org/schema/eml-gbif-profile/1.3/} on
2024-09-25.
}
\examples{
\dontrun{
# check a file
check_eml(file = "https://collections.ala.org.au/ws/eml/dr368")
}
}
