% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm-package.R
\docType{package}
\name{elm-package}
\alias{elm}
\alias{elm-package}
\title{Convert ecological metadata from markdown to xml}
\description{
\code{{elm}} is in development. It is similar to \code{{emld}}, but with the goal of
seamlessly converting between markdown and R formats such as tibbles, lists,
and xml.

'elm' is the common name for the genus 'Ulmus', which is a common tree in the
Northern Hemisphere. It is also an anagram of 'EML', which is an acronym for
Ecological Meta Language, the format that this package creates.
}
\section{Functions}{

\strong{Reading and writing}
\itemize{
\item \code{\link[=read_md_chr]{read_md_chr()}} import markdown as tibble
\item \code{\link[=read_md_xml]{read_md_xml()}} import xml as tibble
\item \code{\link[=write_md_chr]{write_md_chr()}} write \code{elm} objects to markdown files
\item \code{\link[=write_md_xml]{write_md_xml()}} write \code{elm} objects to xml files
}

\strong{Type Conversion}
\itemize{
\item \code{\link[=as_md_chr]{as_md_chr()}} - convert to class \code{md_chr}
\item \code{\link[=as_md_tibble]{as_md_tibble()}} - convert to class \code{md_tibble}
\item \code{\link[=as_md_list]{as_md_list()}} - convert to class \code{md_list}
\item \code{\link[=as_md_xml]{as_md_xml()}} - convert to class \code{md_xml}
\item \code{\link[=add_eml_row]{add_eml_row()}} - add an 'EML' row to an object of class \code{md_tibble}
}

\strong{Parsers}
\itemize{
\item \code{\link[=parse_chr_to_tibble]{parse_chr_to_tibble()}}
\item \code{\link[=parse_tibble_to_list]{parse_tibble_to_list()}}
\item \code{\link[=parse_list_to_xml]{parse_list_to_xml()}}
\item \code{\link[=parse_xml_to_list]{parse_xml_to_list()}}
\item \code{\link[=parse_list_to_tibble]{parse_list_to_tibble()}}
\item \code{\link[=parse_tibble_to_chr]{parse_tibble_to_chr()}}
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
