% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\title{Example metadata statement}
\format{
\subsection{\code{example_metadata}}{

A tibble with X rows and 4 columns, where each row is an xml heading:
\describe{
\item{level}{The nestedness level}
\item{label}{The xml tag}
\item{text}{Text content}
\item{attributes}{list-column containing xml attributes}
}
}
}
\source{
ALA
}
\usage{
example_metadata
}
\description{
This is a very minimalist metadata statement, stored as a tibble, that you
can adapt for your own purposes.
}
\keyword{datasets}
