% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-methods.R
\name{md-methods}
\alias{md-methods}
\alias{print.md}
\alias{summary.md}
\title{Methods for class \code{md}}
\usage{
\method{print}{md}(x, n = 10, ...)

\method{summary}{md}(x, ...)
}
\arguments{
\item{x}{An object of class \code{md}}

\item{n}{(integer) The number of rows of content to print. Defaults to 10.}
}
\description{
These are pretty basic, but should be useful.
}
