% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_chr-methods.R
\name{md_chr-methods}
\alias{md_chr-methods}
\alias{print.md_chr}
\alias{summary.md_chr}
\title{Methods for class \code{md_chr}}
\usage{
\method{print}{md_chr}(x, n = 10, ...)

\method{summary}{md_chr}(object, ...)
}
\arguments{
\item{x}{An object of class \code{md_chr}}

\item{n}{(integer) The number of rows of content to print. Defaults to 10.}

\item{...}{Arguments passed to other methods. Currently ignored.}

\item{object}{An \code{.md} object to summarize}
}
\description{
These are pretty basic, but should be useful.
}
