% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_example}
\alias{metadata_example}
\title{Example metadata statement}
\format{
\subsection{\code{metadata_example}}{

A tibble with X rows and 4 columns, where each row is an xml heading:
\describe{
\item{level}{integer: Nestedness level}
\item{label}{character: xml tag}
\item{text}{list-colum: may contain \code{NA} for no content, or any type of vector
that can be manipulated to a string. Where the entry is a list, list-entries
are treated as separate paragraphs.}
\item{attributes}{list-column: xml attributes. Entries should either be \code{NA}
or a named list.}
}
}
}
\source{
ALA
}
\usage{
metadata_example
}
\description{
This is a very minimalist metadata statement, stored as a tibble, that you
can adapt for your own purposes.
}
\keyword{datasets}
