% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperbark-package.R
\docType{package}
\name{paperbark-package}
\alias{paperbark}
\alias{paperbark-package}
\title{Convert ecological metadata from markdown to xml}
\description{
\code{{paperbark}} is a package to convert markdown documents to xml and back.
}
\section{Functions}{

\strong{Top-level functions}
\itemize{
\item \code{\link[=read_eml]{read_eml()}} - read EML from an \code{xml} file to a \code{tibble}
\item \code{\link[=read_md]{read_md()}} - read metadata from a markdown file to a \code{tibble}
\item \code{\link[=write_eml]{write_eml()}} - write an object to \code{xml}
\item \code{\link[=write_md]{write_md()}} - write an object to markdown
\item \code{\link[=add_eml_header]{add_eml_header()}} - add an 'EML' header row to a \code{tibble}
\item \code{\link[=check_eml]{check_eml()}} - run checks of a file or object against relevant standards
}

\strong{Intermediate-level functions}
\itemize{
\item \code{\link[=as_eml_chr]{as_eml_chr()}} - convert to class \code{character}
\item \code{\link[=as_eml_tibble]{as_eml_tibble()}} - convert to class \code{tbl_df} (i.e. a tibble)
\item \code{\link[=as_eml_list]{as_eml_list()}} - convert to class \code{list}
\item \code{\link[=as_eml_xml]{as_eml_xml()}} - convert to class \code{xml_document}
}

\strong{Low-level functions}
\itemize{
\item \code{\link[=parse_chr_to_tibble]{parse_chr_to_tibble()}} - convert from \code{character} to \code{tibble}
\item \code{\link[=parse_tibble_to_list]{parse_tibble_to_list()}} - convert from \code{tibble} to \code{list}
\item \code{\link[=parse_list_to_xml]{parse_list_to_xml()}} - convert from \code{list} to \code{xml_document}
\item \code{\link[=parse_xml_to_list]{parse_xml_to_list()}} - convert from \code{xml_document} to \code{list}
\item \code{\link[=parse_list_to_tibble]{parse_list_to_tibble()}} - convert from \code{list} to \code{tibble}
\item \code{\link[=parse_tibble_to_chr]{parse_tibble_to_chr()}} - convert from \code{tibble} to \code{character}
}

\strong{Example data}
\itemize{
\item \link{metadata_example} - a boilerplate metadata statement
\item \code{\link[=use_metadata]{use_metadata()}} - create a metadata statement as a markdown file
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramaniam \email{shandiya.balasubramaniam@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
