% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chr_to_tibble.R,
%   R/parse_tibble_to_list.R, R/parse_list_to_xml.R, R/parse_xml_to_list.R,
%   R/parse_list_to_tibble.R, R/parse_tibble_to_chr.R
\name{parse_}
\alias{parse_}
\alias{parse_chr_to_tibble}
\alias{parse_tibble_to_list}
\alias{parse_list_to_xml}
\alias{parse_xml_to_list}
\alias{parse_list_to_tibble}
\alias{parse_tibble_to_chr}
\title{Parse objects from one class to another}
\usage{
parse_chr_to_tibble(x)

parse_tibble_to_list(x)

parse_list_to_xml(x)

parse_xml_to_list(x)

parse_list_to_tibble(x)

parse_tibble_to_chr(x)
}
\arguments{
\item{x}{an R object of the requisite type. No type checking is done.}
}
\description{
These low-level functions are primarily built to be called by the various
\code{as_md} functions, but are exported for greater transparency and bug-fixing.
}
