% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_elm.R
\name{read_elm}
\alias{read_elm}
\alias{read_elm_chr}
\alias{read_elm_xml}
\title{Read metadata from markdown or xml}
\usage{
read_elm(file, type)

read_elm_chr(file)

read_elm_xml(file)
}
\arguments{
\item{file}{filename}
}
\value{
An object of class \code{tbl_df}, \code{tbl} and \code{data.frame} (i.e. a \code{tibble}).
}
\description{
Import metadata from a markdown file, or xml file locally or online.
}
