% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eml.R
\name{read_eml}
\alias{read_eml}
\title{Read an EML-formatted metadata document}
\usage{
read_eml(file)
}
\arguments{
\item{file}{Filename or URL to read from.}
}
\value{
\code{read_eml()} returns an object of class \code{tbl_df}, \code{tbl} and
\code{data.frame} (i.e. a \code{tibble}).
}
\description{
\code{read_eml()} imports metadata from an EML file into the workspace as a
\code{tibble}.
}
\examples{
source_file <- system.file("extdata", 
                           "bionet_metadata.xml",
                           package = "delma")
df <- read_eml(source_file)
}
