% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_md.R
\name{read_md}
\alias{read_md}
\alias{write_md}
\title{Read or write markdown-formatted metadata}
\usage{
read_md(file)

write_md(x, file)
}
\arguments{
\item{file}{Filename to read from or write to. Must be either \code{.md}, \code{.Rmd}
or \code{.Qmd} file.}

\item{x}{Object of any class handled by \code{paperbark}; i.e. \code{character},
\code{tbl_df}, \code{list} or \code{xml_document}.}
}
\value{
\code{read_md()} returns an object of class \code{tbl_df}, \code{tbl} and
\code{data.frame} (i.e. a \code{tibble}). \code{write_md()} doesn't return anything, and
is called for the side-effect of writing the specified markdown file to disk.
}
\description{
\code{read_md()} imports metadata from a markdown file into the workspace as a
\code{tibble}. \code{write_md()} write a \code{string}, \code{tibble}, \code{list} or \code{xml_document}
to a markdown file.
}
\examples{
\dontrun{
# read from a url
file <- system.file("example_data", 
                    "README_md_example.md", 
                    package = "paperbark")
df <- read_md(file)

# write to disk
write_md(df, "example.md")
}
}
