% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_md_xml.R
\name{read_md_xml}
\alias{read_md_xml}
\alias{write_md_xml}
\title{Read or write metadata as \code{xml}}
\usage{
read_md_xml(file)

write_md_xml(x, file)
}
\arguments{
\item{file}{filename}

\item{x}{Object of any class defined by \code{elm}; i.e. \code{md_chr}, \code{md_tibble},
\code{md_list} or \code{md_xml}.}
}
\value{
An object of class \code{md_tibble}.
as defined in \code{xml2}.
}
\description{
Import metadata to or from \code{xml}.  Note that imports to a \code{tibble},
but can export from all \code{elm} formats.
}
