% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{read_xml}
\alias{write_xml}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{xml2}{\code{\link[xml2]{read_xml}}, \code{\link[xml2]{write_xml}}}
}}

