% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_metadata.R
\name{render_metadata}
\alias{render_metadata}
\title{Render a metadata statement to EML}
\usage{
render_metadata(
  input,
  output_file = NULL,
  output_dir = NULL,
  overwrite = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{input}{A file to be rendered to EML}

\item{output_file}{The name of the output file}

\item{output_dir}{The output directory for the rendered \code{output_file}}

\item{overwrite}{(logical) Should any existing file be overwritten? Defaults
to \code{TRUE}.}

\item{quiet}{(logical) Should messages be suppressed? Defaults to \code{FALSE}.}
}
\value{
Does not return an object; called for the side-effect of rendering
a file to EML.
}
\description{
Basically an expanded version of \link[rmarkdown:render]{rmarkdown::render}/\link[quarto:quarto_render]{quarto::quarto_render}
to convert a metadata statement to EML. Note that this process ignores
the \code{output} section of the supplied \code{yaml}, though that should still be
supplied for rendering via other packages. Internally this function wraps
\code{\link[=read_md]{read_md()}} and \code{\link[=write_eml]{write_eml()}} in that order; see those functions for details.
}
