% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_metadata.R
\name{use_metadata}
\alias{use_metadata}
\title{Write an example metadata template to the working directory}
\usage{
use_metadata(file = "metadata.Rmd")
}
\arguments{
\item{file}{(string) a file name to place}
}
\description{
Test function
}
