% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_md_chr.R
\name{as_md_chr}
\alias{as_md_chr}
\alias{as_md_chr.md_tibble}
\alias{as_md_chr.md_list}
\alias{as_md_chr.md_xml}
\title{Convert metadata to class \code{md}}
\usage{
as_md_chr(x, ...)

\method{as_md_chr}{md_tibble}(x, ...)

\method{as_md_chr}{md_list}(x, ...)

\method{as_md_chr}{md_xml}(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Other arguments, currently ignored}
}
\description{
This is basically a vector
}
