% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_md_xml.R
\name{as_md_xml}
\alias{as_md_xml}
\alias{as_md_xml.md_chr}
\alias{as_md_xml.md_tibble}
\alias{as_md_xml.md_list}
\title{Convert metadata to class \code{md_xml}}
\usage{
as_md_xml(x, ...)

\method{as_md_xml}{md_chr}(x, ...)

\method{as_md_xml}{md_tibble}(x, ...)

\method{as_md_xml}{md_list}(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Other arguments, currently ignored}
}
\description{
largely powered by \code{xml2}
}
