% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delma-package.R
\docType{package}
\name{delma-package}
\alias{delma}
\alias{delma-package}
\title{Convert metadata between markdown and EML}
\description{
Ecological Meta Language is a common framework for describing ecological
datasets so they can be shared and reused. \code{delma} supports users to
write metadata statements in markdown for greater transparency and
ease-of-use, then convert them to EML for efficient transfer.
}
\section{Functions}{


\strong{Basic usage}
\itemize{
\item \code{\link[=read_eml]{read_eml()}}/\code{\link[=write_eml]{write_eml()}} - Convert between EML and tibbles
\item \code{\link[=read_md]{read_md()}}/\code{\link[=write_md]{write_md()}} - Convert between markdown files and tibbles
\item \code{\link[=check_eml]{check_eml()}} - Check your metadata statement against relevant standards
}

\strong{Type conversion}
\itemize{
\item \code{\link[=as_eml_chr]{as_eml_chr()}} - Convert metadata to class \code{character}
\item \code{\link[=as_eml_tibble]{as_eml_tibble()}} - Convert metadata to class \code{tbl_df} (i.e. a tibble)
\item \code{\link[=as_eml_list]{as_eml_list()}} - Convert metadata to class \code{list}
\item \code{\link[=as_eml_xml]{as_eml_xml()}} - Convert metadata to class \code{xml_document}
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paperbark.ala.org.au}
}

}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramaniam \email{shandiya.balasubramaniam@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
