% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm-package.R
\docType{package}
\name{elm-package}
\alias{elm}
\alias{elm-package}
\title{Convert ecological metadata from markdown to xml}
\description{
\code{{elm}} is a package to convert markdown documents to xml and back.
'elm' is the common name for the genus 'Ulmus', which is a common tree in the
Northern Hemisphere. It is also an anagram of 'EML', which is an acronym for
Ecological Meta Language, the format that this package creates.
}
\section{Functions}{

\strong{Reading and writing}
\itemize{
\item \code{\link[=read_elm]{read_elm()}} import markdown or \code{xml} to a \code{tibble}
\item \code{\link[=write_elm]{write_elm()}} write an object to markdown or \code{xml}
}

\strong{Type Conversion}
\itemize{
\item \code{\link[=as_elm_chr]{as_elm_chr()}} - convert to class \code{character}
\item \code{\link[=as_elm_tibble]{as_elm_tibble()}} - convert to class \code{tbl_df} (i.e. a tibble)
\item \code{\link[=as_elm_list]{as_elm_list()}} - convert to class \code{list}
\item \code{\link[=as_elm_xml]{as_elm_xml()}} - convert to class \code{xml_document}
\item \code{\link[=add_elm_header]{add_elm_header()}} - add an 'EML' header row to a \code{tibble}
}

\strong{Parsers}
\itemize{
\item \code{\link[=parse_chr_to_tibble]{parse_chr_to_tibble()}}
\item \code{\link[=parse_tibble_to_list]{parse_tibble_to_list()}}
\item \code{\link[=parse_list_to_xml]{parse_list_to_xml()}}
\item \code{\link[=parse_xml_to_list]{parse_xml_to_list()}}
\item \code{\link[=parse_list_to_tibble]{parse_list_to_tibble()}}
\item \code{\link[=parse_tibble_to_chr]{parse_tibble_to_chr()}}
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
