% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm-package.R
\docType{package}
\name{elm-package}
\alias{elm}
\alias{elm-package}
\title{Convert ecological metadata from markdown to xml}
\description{
\code{{elm}} is a package to convert markdown documents to xml and back.
'elm' is the common name for the genus 'Ulmus', which is a common tree in the
Northern Hemisphere. It is also an anagram of 'EML', which is an acronym for
Ecological Meta Language, the format that this package creates.
}
\section{Functions}{

\strong{Example data}
\itemize{
\item \link{metadata_example}
\item \code{\link[=add_eml_header]{add_eml_header()}} - add an 'EML' header row to a \code{tibble}
\item \code{\link[=use_metadata]{use_metadata()}}
\item \code{\link[=check_eml]{check_eml()}}
}

\strong{Low-level functions}
\itemize{
\item \code{\link[=parse_chr_to_tibble]{parse_chr_to_tibble()}}
\item \code{\link[=parse_tibble_to_list]{parse_tibble_to_list()}}
\item \code{\link[=parse_list_to_xml]{parse_list_to_xml()}}
\item \code{\link[=parse_xml_to_list]{parse_xml_to_list()}}
\item \code{\link[=parse_list_to_tibble]{parse_list_to_tibble()}}
\item \code{\link[=parse_tibble_to_chr]{parse_tibble_to_chr()}}
}

\strong{Intermediate-level functions}
\itemize{
\item \code{\link[=as_eml_chr]{as_eml_chr()}} - convert to class \code{character}
\item \code{\link[=as_eml_tibble]{as_eml_tibble()}} - convert to class \code{tbl_df} (i.e. a tibble)
\item \code{\link[=as_eml_list]{as_eml_list()}} - convert to class \code{list}
\item \code{\link[=as_eml_xml]{as_eml_xml()}} - convert to class \code{xml_document}
}

\strong{Top-level functions}
\itemize{
\item \code{\link[=read_eml]{read_eml()}} read EML from an \code{xml} file
\item \code{\link[=read_md]{read_md()}} read a markdown file
\item \code{\link[=write_eml]{write_eml()}} write an object to \code{xml}
\item \code{\link[=write_md]{write_md()}} write an object to markdown
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
