% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elm-package.R
\docType{package}
\name{elm-package}
\alias{elm}
\alias{elm-package}
\title{Convert ecological metadata from markdown to xml}
\description{
\code{{elm}} is in development. It is similar to \code{{emld}}, but with the goal of
seamlessly converting between markdown and R formats such as tibbles, lists,
and xml.

'elm' is the common name for the genus 'Ulmus', which is a common tree in the
Northern Hemisphere. It is also an anagram of 'EML', which is an acronym for
Ecological Meta Language, the format that this package creates.
}
\section{Functions}{

\strong{Reading and writing}
\itemize{
\item \code{\link[=read_md]{read_md()}} to import from Markdown files
\item \code{\link[=read_xml]{read_xml()}} to import from XML - reexported from \code{{xml2}}
}

\strong{Type Conversion}
\itemize{
\item \code{\link[=as_md]{as_md()}} - markdown stored as strings (also \code{\link[=md-methods]{md-methods()}})
\item \code{\link[=as_tibble]{as_tibble()}} - tibbles (from package \code{{tibble}})
\item \code{\link[=as_eml]{as_eml()}} - list-like data
\item \code{\link[=as_xml]{as_xml()}} - xml data
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
