% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eml.R
\name{read_eml}
\alias{read_eml}
\title{Read metadata from xml}
\usage{
read_eml(file)
}
\arguments{
\item{file}{Filename or url}
}
\value{
An object of class \code{tbl_df}, \code{tbl} and \code{data.frame} (i.e. a \code{tibble}).
}
\description{
Import metadata from an xml file locally or online.
}
