% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_md.R
\name{read_md}
\alias{read_md}
\alias{write_md}
\title{Import a markdown file as text}
\usage{
read_md(file)

write_md(x, file)
}
\arguments{
\item{file}{filename}

\item{x}{object of class \code{md}}
}
\description{
Very basic function to create an object of class \code{md} (i.e. a \code{vector} of
class \code{character}).
}
