% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_md.R
\name{read_md}
\alias{read_md}
\title{Read metadata from markdown}
\usage{
read_md(file)
}
\arguments{
\item{file}{A file to import.}
}
\value{
An object of class \code{tbl_df}, \code{tbl} and \code{data.frame} (i.e. a \code{tibble}).
}
\description{
Import metadata from a markdown file.
}
