% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_md_chr.R
\name{read_md_chr}
\alias{read_md_chr}
\alias{write_md_chr}
\title{Handle metadata as markdown}
\usage{
read_md_chr(file)

write_md_chr(x, file)
}
\arguments{
\item{file}{filename}

\item{x}{Object of any class defined by \code{elm}; i.e. \code{md_chr}, \code{md_tibble},
\code{md_list} or \code{md_xml}.}
}
\value{
An object of class \code{md_tibble}.
}
\description{
Import metadata to or from a markdown file. Note that imports to a \code{tibble},
but can export from all \code{elm} formats.
}
