% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_elm.R
\name{write_elm}
\alias{write_elm}
\alias{write_elm_chr}
\alias{write_elm_xml}
\title{Write metadata to markdown or xml}
\usage{
write_elm_chr(x, file)

write_elm_xml(x, file)
}
\arguments{
\item{x}{Object of any class handled by \code{elm}; i.e. \code{character}, \code{tbl_df},
\code{list} or \code{xml_document}.}

\item{file}{filename to save to}
}
\description{
Write metadata from a \code{tibble}, \code{character}, \code{list} or \code{xml_document} to
a markdown (\code{.md}) or xml (\code{.xml}) document.
}
