% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_eml.R
\name{write_eml}
\alias{write_eml}
\title{Write metadata to xml}
\usage{
write_eml(x, file)
}
\arguments{
\item{x}{Object of any class handled by \code{elm}; i.e. \code{character}, \code{tbl_df},
\code{list} or \code{xml_document}.}

\item{file}{Filename to export to.}
}
\description{
Write metadata from a \code{tibble}, \code{character}, \code{list} or \code{xml_document} to
an xml (\code{.xml}) document.
}
