% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_md.R
\name{write_md}
\alias{write_md}
\title{Write metadata to markdown}
\usage{
write_md(x, file)
}
\arguments{
\item{x}{Object of any class handled by \code{paperbark}; i.e. \code{character},
\code{tbl_df}, \code{list} or \code{xml_document}.}

\item{file}{Filename to export to.}
}
\description{
Write metadata from a \code{tibble}, \code{character}, \code{list} or \code{xml_document} to
a markdown (\code{.md}) document.
}
