% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_md.R
\name{write_md}
\alias{write_md}
\title{Write a markdown-formatted metadata document}
\usage{
write_md(x, file)
}
\arguments{
\item{x}{Object of any class handled by \code{delma}; i.e. \code{tbl_df}, \code{list} or
\code{xml_document}.
#' @param file Filename to write to. Must be either \code{.md}, \code{.Rmd}
or \code{.Qmd} file.}
}
\description{
Write a markdown-formatted metadata document
}
\details{
Similar to \code{\link[=read_md]{read_md()}}, \code{\link[=write_md]{write_md()}} is considerably less generic than most
\code{write_} functions. If \code{x} is an \code{xml_document} this should convert seamlessly;
but lists or tibbles that have been manually formatted require care. To parse
correctly, the supplied \code{tibble} \strong{must} contain the columns supplied by \code{\link[=read_md]{read_md()}}.

Internally, \code{write_md()} calls \link[lightparser:combine_tbl_to_file]{lightparser::combine_tbl_to_file}.
}
